Use instant scroll-to-top for links by default
